requirements.txt
Flask==2.3.2
WTForms==3.0.1
Werkzeug==2.3.6
Flask-SQLAlchemy==3.0.5
Flask-Admin==1.6.1
Flask-Login==0.6.2
gunicorn>=20.1.0
Flask-QRcode==3.1.0
python-barcode==0.15.1
Flask-Bcrypt==1.0.1